// src/models/Register.ts
/**
 *
 * @export
 * @interface Register
 */
export interface Register {
    /**
     *
     * @type {string}
     * @memberof Register
     */
    username: string;
    /**
     *
     * @type {string}
     * @memberof Register
     */
    lastName: string;
    /**
     *
     * @type {string}
     * @memberof Register
     */
    mail: string;
    /**
     *
     * @type {string}
     * @memberof Register
     */
    password: string;
    /**
     *
     * @type {string}
     * @memberof Register
     */
    confirmPass: string;
  
  }
  