// src/models/User.ts
import internal from "stream";

/**
 *
 * @export
 * @interface User
 */
export interface User {
    /**
     *
     * @type {string}
     * @memberof User
     */
    Correo: string;
    /**
     *
     * @type {string}
     * @memberof User
     */
    Contrasena: string;
    /**
     *
     * @type {string}
     * @memberof User
     */
    Nombres: string;
    /**
     *
     * @type {string}
     * @memberof User
     */
    Apellidos: string;
    /**
     *
     * @type {number}
     * @memberof User
     */
    Role: number;
  
  }
  